LayoutTests/accessibility/mac/aria-menuitem-checked-value-expected.txt

This tests that the right Mac attribute for conveying 'checked' states works on menu items

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Item 1 Mark: ✓
PASS !item2mark || item2mark == '' is true
Item 3 Mark: ✓
PASS !item4mark || item4mark == '' is true
PASS successfullyParsed is true

TEST COMPLETE
